+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Added
+* `Equal` and `EqualCounts`
 
 ## [0.3.0](https://github.com/coady/iterset/releases/tag/v0.3.0) - 2025-01-26
 ### Added